Move ternary button setup to `HitObjectComposer`
As per https://github.com/ppy/osu/pull/37848#discussion_r3287018539,
this PR migrates the code responsible for creating the ternary buttons
used for the editor toggles and sample bank from `ComposeBlueprintContainer`
to `HitObjectComposer`, mainly because it has nothing to do with blueprints.

Also it makes the new combo icon added in https://github.com/ppy/osu/pull/37848
automatically generated based on the first tool in the toolbox.
