+++ osu.Game.Rulesets.Catch/Edit/CatchBlueprintContainer.cs
@@ -3,16 +3,11 @@
 
 using System.Collections.Generic;
 using System.Linq;
-using osu.Framework.Graphics;
-using osu.Framework.Graphics.Containers;
-using osu.Framework.Graphics.Sprites;
 using osu.Framework.Input.Events;
-using osu.Game.Graphics;
 using osu.Game.Rulesets.Catch.Edit.Blueprints;
 using osu.Game.Rulesets.Catch.Objects;
 using osu.Game.Rulesets.Edit;
 using osu.Game.Rulesets.Objects;
-using osu.Game.Screens.Edit.Components.TernaryButtons;
 using osu.Game.Screens.Edit.Compose.Components;
 using osuTK;
 
@@ -26,29 +21,6 @@ namespace osu.Game.Rulesets.Catch.Edit
             : base(composer)
         {
         }
-
-        protected override Drawable CreateNewComboButton() => new NewComboTernaryButton
-        {
-            Current = NewCombo,
-            CreateIcon = () => new Container
-            {
-                Children = new Drawable[]
-                {
-                    new SpriteIcon
-                    {
-                        Anchor = Anchor.BottomLeft,
-                        Origin = Anchor.BottomLeft,
-                        Icon = OsuIcon.EditorFruit,
-                        Size = new Vector2(15),
-                    },
-                    new SpriteIcon
-                    {
-                        Icon = OsuIcon.EditorNewComboSparkles,
-                        Size = new Vector2(20),
-                    }
-                },
-            },
-        };
 
         protected override SelectionHandler<HitObject> CreateSelectionHandler() => new CatchSelectionHandler();
